**Refactor AI settings handling and improve configuration management**
- Simplified AI settings opening logic by introducing `llm::Manager::OpenSettingsFileInEditor()` which handles file creation, backup, and version checking
- Removed redundant global settings retrieval and file handling code from `ChatAIWindow`
- Updated `LLMManager` to include versioning support for configuration files
- Modified `clStatusBar` to use `size_t` instead of `int` for field index parameters for better type safety
- Removed unused UI elements and event bindings related to settings in `ChatAIWindow`

+++ Plugin/clStatusBar.cpp
@@ -286,7 +286,7 @@ void clStatusBar::StopAnimation()
     field->SetTooltip("");
 }
 
-void clStatusBar::OnFieldClicked(clCommandEvent& event) { DoFieldClicked(event.GetInt()); }
+void clStatusBar::OnFieldClicked(clCommandEvent& event) { DoFieldClicked(static_cast<size_t>(event.GetInt())); }
 
 void clStatusBar::SetWhitespaceInfo()
 {
@@ -388,7 +388,7 @@ void clStatusBar::DoUpdateView()
     SetWhitespaceInfo();
 }
 
-void clStatusBar::DoFieldClicked(int fieldIndex)
+void clStatusBar::DoFieldClicked(size_t fieldIndex)
 {
     if (fieldIndex == STATUSBAR_SCM_IDX) {
         if (m_sourceControlTabName.IsEmpty())
